Extract grouping columns verifications to MailchimpClient
